requirements.txt
# Project
-f https://www.djangoproject.com/m/releases/1.4/Django-1.4c1.tar.gz Django==1.4c1
South==0.7.3
psycopg2==2.4.3
wsgiref==0.1.2
django-model-utils==1.0.0
feedparser==5.1
pytz==2011n
-e git+https://github.com/alex/django-filter.git@51b39fca089f2cb1d3d3501d2e164635fbe52864#egg=django_filter-dev
-e git://github.com/earle/django-bootstrap.git@484837e9cfcae64ba6f0ee06acbd547648f0cca8#egg=bootstrap-dev
django-taggit==0.9.3
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1

# Development
django-debug-toolbar==0.8.5

# Documentation
Jinja2==2.6
Pygments==1.4
Sphinx==1.0.7
docutils==0.8.1

# Deployment
Fabric==1.3.3
pycrypto==2.3
ssh==1.7.11
littlechef==1.0.4
paramiko==1.7.7.1
simplejson==2.3.0

# Google Docs
gdata==2.0.16
